% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_who_age_groups}
\alias{add_who_age_groups}
\title{Add WHO conforming age groups naively}
\usage{
add_who_age_groups(.data, is_who)
}
\arguments{
\item{.data}{a tibble with 'age_group' variable}

\item{is_who}{T/F, are we using who?}
}
\value{
The dataset with maybe one more variable (who_age_group)
}
\description{
Add WHO conforming age groups naively
}
