% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahp.R
\name{mahp}
\alias{mahp}
\title{A model of alchohol harms and policies}
\description{
API for InterMAHP Shiny App
}
\section{Fields}{

\describe{
 \item{pc}{A dataset of alcohol consumption and prevalence}
 \item{mm}{A dataset of mortality and morbidity counts}
 \item{rr}{A dataset of relative risk function evaluations}
 \item{sk}{A dataset skeleton used to perform aaf computations}
 \item{af}{A long dataset of attributable fractions}
 \item{dg}{A list of gender-stratified drinking groups}
 \item{bb}{Gender-stratified definition of binge drinking}
 \item{scc}{Gender-stratified propotion of squamous-cell carcinoma among
   oesophageal cancers (only SCC is alcohol-caused)}
 \item{ub}{Upper bound of alcohol consumption}
 \item{dg}{List of drinking groups (gender stratified)}
 \item{ext}{Boolean indicating whether relative risk functions are
   extrapolated linearly (TRUE) or capped (FALSE) after a consumption level
   of 150 grams per day}
 \item{rr_str}{String indicating which relative risk function source we use.
   his is used when the rr dataset needs updating via update_rr()}
}
}

\section{Data Methods}{

\code{$new()} Creates a new model object.

\code{$add_pc()} Prepares a prevalence and consumption dataset

\code{$add_mm()} Prepares a morbidity and mortality dataset

\code{$choose_rr()} Chooses a source for relative risk functions

\code{$choose_project()} Chooses a predefined set of project settings

\code{$set_ext()} Set risk extrapolation method (linear >> TRUE, capped >> FALSE)

\code{$set_bb()} Sets binge consumption definitions

\code{$set_ub()} Sets upper bound on consumption

\code{$set_scc()} Sets squamous cell carcinoma proportions

\code{$update_rr()} Updates stored rr dataset
}

\section{Evaluation Methods}{

\code{$init_sk()} Initializes the skeleton computation dataset and evaluates
 at baseline consumption

\code{$add_scenario()} Adds new scenario attributable fractions and relative
 attributable fractions to the skeleton dataset for each existing drinking
 group

\code{$add_group()} Adds a new set of drinking groups for each existing
 scenario
}

