% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{screen_gender}
\alias{screen_gender}
\title{Screens supplied gender variable for InterMAHP permissable values}
\usage{
screen_gender(.char)
}
\description{
Due to the availability of dose-response relative risk functions, we require
the gender variable to have one of 2 values: w or m. In the frontend, we will
include an interactive component for distributing uploaded values for gender.
}
