% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{screen_mm}
\alias{screen_mm}
\title{Screens a morbidity and mortality dataset for problematic measures}
\usage{
screen_mm(.data)
}
\description{
Screens a morbidity and mortality dataset for problematic measures
}
