% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{screen_pc}
\alias{screen_pc}
\title{Screens a prevalence and consumption dataset for problematic measures}
\usage{
screen_pc(.data)
}
\description{
Screens a prevalence and consumption dataset for problematic measures
}
