% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{screen_vars}
\alias{screen_vars}
\title{Lowers case and checks for missing/extra variables}
\usage{
screen_vars(.data, .vars, .name)
}
\arguments{
\item{.vars}{A character vector of variable names, all lower case.}
}
\description{
Lowers case and checks for missing/extra variables
}
