% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahp.R
\name{mahp}
\alias{mahp}
\title{A model of alchohol harms and policies}
\description{
API for InterMAHP Shiny App
}
\section{Input Fields}{

\describe{
 \item{pc}{A dataset of alcohol consumption and prevalence}
 \item{mm}{A dataset of mortality and morbidity counts}
 \item{rr}{A list of relative risk function evaluation datasets}
 \item{sk}{A dataset skeleton used to perform aaf computations}
 \item{dg}{A list of gender-stratified drinking groups}
 \item{bb}{Gender-stratified definition of binge drinking}
 \item{scc}{Gender-stratified propotion of squamous-cell carcinoma among
   oesophageal cancers (only SCC is alcohol-caused)}
 \item{ub}{Upper bound of alcohol consumption}
 \item{mcn}{Number of samples in Monte Carlo uncertainty estimation}
 \item{pc_sample_vars}{Which prevalence and consumption variables to sample}
 \item{dg}{List of all drinking groups (gender stratified)}
 \item{sn}{Vector of all scenarios as mult. changes in consumption.}
 \item{ext}{Character indicating whether relative risk functions are
   extrapolated linearly or capped after a consumption level
   of 150 grams per day}
 \item{cal}{Boolean indiciating whether to try to calibrate absolute risk
   curves for calibrable wholly attributable conditions}
 \item{rr_choice}{Character, relative risk source}
}
}

\section{Computed Fields}{

\describe{
 \item{af}{A list of the wide fraction datasets described below}
 \item{base_paf}{A wide dataset of alcohol attributable fractions from
   partially attributable causes not affected by bingeing}
 \item{base_former_paf}{A wide dataset of alcohol attributable fractions from
   partially attributable causes not affected by bingeing that still affect
   former drinkers}
 \item{binge_paf}{A wide dataset of alcohol attributable fractions from
   partially attributable causes affected by bingeing}
 \item{binge_former_paf}{A wide dataset of alcohol attributable fractions
   from partially attributable causes affected by bingeing that still affect
   former drinkers}
 \item{scaled_base_waf}{A wide dataset of alcohol attributable fractions from
   wholly attributable causes not affected by binging whose attributable
   fractions are scaled forms of similar partially attributable forms}
 \item{scaled_binge_waf}{A wide dataset of alcohol attributable fractions
   from wholly attributable causes affected by binging whose attributable
   fractions are scaled forms of similar partially attributable forms}
 \item{calibrated_waf}{A wide dataset of alcohol attributable fractions from
   wholly attributable causes whose risk functions are calibrated from
   morbidity and mortality data}
}
}

\section{Data Methods}{

\code{$new()} Creates a new model object.

\code{$add_pc()} Prepares a prevalence and consumption dataset

\code{$add_mm()} Prepares a morbidity and mortality dataset

\code{$choose_rr()} Chooses a source for relative risk functions

\code{$choose_project()} Chooses a predefined set of project settings

\code{$set_ext()} Set risk extrapolation method (linear >> TRUE, capped >> FALSE)

\code{$set_bb()} Sets binge consumption definitions

\code{$set_ub()} Sets upper bound on consumption

\code{$set_scc()} Sets squamous cell carcinoma proportions

\code{$update_rr()} Updates relative risk function evaluations with latest
 parameters

\code{$make_gamma()} Makes base and binge gamma functions at the prescribed
 level of consumption
}

\section{Evaluation Methods}{

\code{$init_fractions} Initialize and populate fraction sheets

\code{$def_scenario()} Defines a scenario
\code{$rm_scenario}
\code{$cmp_scenario()} Adds new scenario attributable fractions and relative attributable fractions to the partially attributable fraction dataset
\code{$cmp_scenarios()}

\code{$def_group()} Defines or redefines a group
\code{$rm_group}
\code{$cmp_group()}
\code{$cmp_groups()}


\code{$add_scenario()} Adds new scenario attributable fractions and relative
 attributable fractions to the skeleton dataset for each existing drinking
 group

\code{$add_group()} Adds a new set of drinking groups for each existing
 scenario
}

\section{Results Methods}{

\code{$get_afs} Provides all attributable fractions computed and evaluates scc
 correction if supplied

\code{$get_long_afs} Invokes get_afs and formats as long

\code{$get_long_counts} Invokes get_long_afs and applied afs to counts
}

